+++ spec/ductwork/configuration/job_worker_count_spec.rb
@@ -22,6 +22,14 @@ RSpec.describe Ductwork::Configuration, "#job_worker_count" do
 
       expect(config.job_worker_count("foobar")).to eq(5)
     end
+
+    it "returns the manually set value" do
+      create_default_config_file
+      config = described_class.new
+      config.job_worker_count = 1_000
+
+      expect(config.job_worker_count("foo")).to eq(1_000)
+    end
   end
 
   context "with a count for the specific pipeline" do